Network.AWS.Athena.BatchGetQueryExecution

Description

Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings. To get a list of query execution IDs, use ListQueryExecutions . Query executions are different from named (saved) queries. Use BatchGetNamedQuery to get details about named queries.

Creating a Request

batchGetQueryExecution

Arguments

:: NonEmpty Text
bgqeQueryExecutionIds

-> BatchGetQueryExecution

Creates a value of BatchGetQueryExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgqeQueryExecutionIds - An array of query execution IDs.

data BatchGetQueryExecution

See: batchGetQueryExecution smart constructor.

Request Lenses

bgqeQueryExecutionIds :: Lens' BatchGetQueryExecution (NonEmpty Text)

An array of query execution IDs.

Destructuring the Response

batchGetQueryExecutionResponse

Arguments

:: Int
bgqersResponseStatus

-> BatchGetQueryExecutionResponse

Creates a value of BatchGetQueryExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgqersUnprocessedQueryExecutionIds - Information about the query executions that failed to run.

bgqersQueryExecutions - Information about a query execution.

bgqersResponseStatus - -- | The response status code.

data BatchGetQueryExecutionResponse

See: batchGetQueryExecutionResponse smart constructor.

Response Lenses

bgqersUnprocessedQueryExecutionIds :: Lens' BatchGetQueryExecutionResponse [UnprocessedQueryExecutionId]

Information about the query executions that failed to run.

bgqersQueryExecutions :: Lens' BatchGetQueryExecutionResponse [QueryExecution]

Information about a query execution.

bgqersResponseStatus :: Lens' BatchGetQueryExecutionResponse Int

- | The response status code.