Network.AWS.Athena.CreateNamedQuery

Description

Creates a named query.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide .

Creating a Request

createNamedQuery

Arguments

:: Text
cnqName

-> Text
cnqDatabase

-> Text
cnqQueryString

-> CreateNamedQuery

Creates a value of CreateNamedQuery with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnqClientRequestToken - A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another CreateNamedQuery request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString , an error is returned. Important: This token is listed as not required because AWS SDKs (for example the AWS SDK for Java) auto-generate the token for users. If you are not using the AWS SDK or the AWS CLI, you must provide this token or the action will fail.

cnqDescription - A brief explanation of the query.

cnqName - The plain language name for the query.

cnqDatabase - The database to which the query belongs.

cnqQueryString - The text of the query itself. In other words, all query statements.

data CreateNamedQuery

See: createNamedQuery smart constructor.

Request Lenses

cnqClientRequestToken :: Lens' CreateNamedQuery (Maybe Text)

A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another CreateNamedQuery request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString , an error is returned. Important: This token is listed as not required because AWS SDKs (for example the AWS SDK for Java) auto-generate the token for users. If you are not using the AWS SDK or the AWS CLI, you must provide this token or the action will fail.

cnqDescription :: Lens' CreateNamedQuery (Maybe Text)

A brief explanation of the query.

cnqName :: Lens' CreateNamedQuery Text

The plain language name for the query.

cnqDatabase :: Lens' CreateNamedQuery Text

The database to which the query belongs.

cnqQueryString :: Lens' CreateNamedQuery Text

The text of the query itself. In other words, all query statements.

Destructuring the Response

createNamedQueryResponse

Arguments

:: Int
cnqrsResponseStatus

-> CreateNamedQueryResponse

Creates a value of CreateNamedQueryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnqrsNamedQueryId - The unique ID of the query.

cnqrsResponseStatus - -- | The response status code.

data CreateNamedQueryResponse

See: createNamedQueryResponse smart constructor.

Response Lenses

cnqrsNamedQueryId :: Lens' CreateNamedQueryResponse (Maybe Text)

The unique ID of the query.

cnqrsResponseStatus :: Lens' CreateNamedQueryResponse Int

- | The response status code.