Network.AWS.Athena.ListNamedQueries

Description

Provides a list of all available query IDs.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide .

This operation returns paginated results.

Creating a Request

listNamedQueries :: ListNamedQueries

Creates a value of ListNamedQueries with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lnqNextToken - The token that specifies where to start pagination if a previous request was truncated.

lnqMaxResults - The maximum number of queries to return in this request.

data ListNamedQueries

See: listNamedQueries smart constructor.

Request Lenses

lnqNextToken :: Lens' ListNamedQueries (Maybe Text)

The token that specifies where to start pagination if a previous request was truncated.

lnqMaxResults :: Lens' ListNamedQueries (Maybe Natural)

The maximum number of queries to return in this request.

Destructuring the Response

listNamedQueriesResponse

Arguments

:: Int
lnqrsResponseStatus

-> ListNamedQueriesResponse

Creates a value of ListNamedQueriesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lnqrsNextToken - A token to be used by the next request if this request is truncated.

lnqrsNamedQueryIds - The list of unique query IDs.

lnqrsResponseStatus - -- | The response status code.

data ListNamedQueriesResponse

See: listNamedQueriesResponse smart constructor.

Response Lenses

lnqrsNextToken :: Lens' ListNamedQueriesResponse (Maybe Text)

A token to be used by the next request if this request is truncated.

lnqrsNamedQueryIds :: Lens' ListNamedQueriesResponse (Maybe (NonEmpty Text))

The list of unique query IDs.

lnqrsResponseStatus :: Lens' ListNamedQueriesResponse Int

- | The response status code.