Network.AWS.Athena.ListQueryExecutions

Description

Provides a list of all available query execution IDs.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide .

This operation returns paginated results.

Creating a Request

listQueryExecutions :: ListQueryExecutions

Creates a value of ListQueryExecutions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lqeNextToken - The token that specifies where to start pagination if a previous request was truncated.

lqeMaxResults - The maximum number of query executions to return in this request.

data ListQueryExecutions

See: listQueryExecutions smart constructor.

Request Lenses

lqeNextToken :: Lens' ListQueryExecutions (Maybe Text)

The token that specifies where to start pagination if a previous request was truncated.

lqeMaxResults :: Lens' ListQueryExecutions (Maybe Natural)

The maximum number of query executions to return in this request.

Destructuring the Response

listQueryExecutionsResponse

Arguments

:: Int
lqersResponseStatus

-> ListQueryExecutionsResponse

Creates a value of ListQueryExecutionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lqersQueryExecutionIds - The unique IDs of each query execution as an array of strings.

lqersNextToken - A token to be used by the next request if this request is truncated.

lqersResponseStatus - -- | The response status code.

data ListQueryExecutionsResponse

See: listQueryExecutionsResponse smart constructor.

Response Lenses

lqersQueryExecutionIds :: Lens' ListQueryExecutionsResponse (Maybe (NonEmpty Text))

The unique IDs of each query execution as an array of strings.

lqersNextToken :: Lens' ListQueryExecutionsResponse (Maybe Text)

A token to be used by the next request if this request is truncated.

lqersResponseStatus :: Lens' ListQueryExecutionsResponse Int

- | The response status code.