Network.AWS.Athena.Types

Description

Service Configuration

athena :: Service

API version 2017-05-18 of the Amazon Athena SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that something is wrong with the input to the request. For example, a required parameter may be missing or out of range.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the request was throttled.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates a platform issue, which may be due to a transient condition or outage.

ColumnNullable

data ColumnNullable

Constructors

NotNull

Nullable

Unknown

EncryptionOption

data EncryptionOption

Constructors

CseKMS

SseKMS

SseS3

QueryExecutionState

data QueryExecutionState

Constructors

Cancelled

Failed

Queued

Running

Succeeded

ColumnInfo

data ColumnInfo

Information about the columns in a query execution result.

See: columnInfo smart constructor.

columnInfo

Arguments

:: Text
ciName

-> Text
ciType

-> ColumnInfo

Creates a value of ColumnInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciScale - For DECIMAL data types, specifies the total number of digits in the fractional part of the value. Defaults to 0.

ciPrecision - For DECIMAL data types, specifies the total number of digits, up to 38. For performance reasons, we recommend up to 18 digits.

ciSchemaName - The schema name (database name) to which the query results belong.

ciCatalogName - The catalog to which the query results belong.

ciCaseSensitive - Indicates whether values in the column are case-sensitive.

ciLabel - A column label.

ciTableName - The table name for the query results.

ciNullable - Indicates the column's nullable status.

ciName - The name of the column.

ciType - The data type of the column.

ciScale :: Lens' ColumnInfo (Maybe Int)

For DECIMAL data types, specifies the total number of digits in the fractional part of the value. Defaults to 0.

ciPrecision :: Lens' ColumnInfo (Maybe Int)

For DECIMAL data types, specifies the total number of digits, up to 38. For performance reasons, we recommend up to 18 digits.

ciSchemaName :: Lens' ColumnInfo (Maybe Text)

The schema name (database name) to which the query results belong.

ciCatalogName :: Lens' ColumnInfo (Maybe Text)

The catalog to which the query results belong.

ciCaseSensitive :: Lens' ColumnInfo (Maybe Bool)

Indicates whether values in the column are case-sensitive.

ciLabel :: Lens' ColumnInfo (Maybe Text)

A column label.

ciTableName :: Lens' ColumnInfo (Maybe Text)

The table name for the query results.

ciNullable :: Lens' ColumnInfo (Maybe ColumnNullable)

Indicates the column's nullable status.

ciName :: Lens' ColumnInfo Text

The name of the column.

ciType :: Lens' ColumnInfo Text

The data type of the column.

Datum

data Datum

A piece of data (a field in the table).

See: datum smart constructor.

datum :: Datum

Creates a value of Datum with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dVarCharValue - The value of the datum.

dVarCharValue :: Lens' Datum (Maybe Text)

The value of the datum.

EncryptionConfiguration

data EncryptionConfiguration

If query results are encrypted in Amazon S3, indicates the Amazon S3 encryption option used.

See: encryptionConfiguration smart constructor.

encryptionConfiguration

Arguments

:: EncryptionOption
ecEncryptionOption

-> EncryptionConfiguration

Creates a value of EncryptionConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ecKMSKey - For SSE-KMS and CSE-KMS , this is the KMS key ARN or ID.

ecEncryptionOption - Indicates whether Amazon S3 server-side encryption with Amazon S3-managed keys (SSE-S3 ), server-side encryption with KMS-managed keys (SSE-KMS ), or client-side encryption with KMS-managed keys (CSE-KMS) is used.

ecKMSKey :: Lens' EncryptionConfiguration (Maybe Text)

For SSE-KMS and CSE-KMS , this is the KMS key ARN or ID.

ecEncryptionOption :: Lens' EncryptionConfiguration EncryptionOption

Indicates whether Amazon S3 server-side encryption with Amazon S3-managed keys (SSE-S3 ), server-side encryption with KMS-managed keys (SSE-KMS ), or client-side encryption with KMS-managed keys (CSE-KMS) is used.

NamedQuery

data NamedQuery

A query, where QueryString is the SQL query statements that comprise the query.

See: namedQuery smart constructor.

namedQuery

Arguments

:: Text
nqName

-> Text
nqDatabase

-> Text
nqQueryString

-> NamedQuery

Creates a value of NamedQuery with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nqNamedQueryId - The unique identifier of the query.

nqDescription - A brief description of the query.

nqName - The plain-language name of the query.

nqDatabase - The database to which the query belongs.

nqQueryString - The SQL query statements that comprise the query.

nqNamedQueryId :: Lens' NamedQuery (Maybe Text)

The unique identifier of the query.

nqDescription :: Lens' NamedQuery (Maybe Text)

A brief description of the query.

nqName :: Lens' NamedQuery Text

The plain-language name of the query.

nqDatabase :: Lens' NamedQuery Text

The database to which the query belongs.

nqQueryString :: Lens' NamedQuery Text

The SQL query statements that comprise the query.

QueryExecution

data QueryExecution

Information about a single instance of a query execution.

See: queryExecution smart constructor.

queryExecution :: QueryExecution

Creates a value of QueryExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qeStatus - The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

qeQueryExecutionContext - The database in which the query execution occurred.

qeResultConfiguration - The location in Amazon S3 where query results were stored and the encryption option, if any, used for query results.

qeQuery - The SQL query statements which the query execution ran.

qeStatistics - The amount of data scanned during the query execution and the amount of time that it took to execute.

qeQueryExecutionId - The unique identifier for each query execution.

qeStatus :: Lens' QueryExecution (Maybe QueryExecutionStatus)

The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

qeQueryExecutionContext :: Lens' QueryExecution (Maybe QueryExecutionContext)

The database in which the query execution occurred.

qeResultConfiguration :: Lens' QueryExecution (Maybe ResultConfiguration)

The location in Amazon S3 where query results were stored and the encryption option, if any, used for query results.

qeQuery :: Lens' QueryExecution (Maybe Text)

The SQL query statements which the query execution ran.

qeStatistics :: Lens' QueryExecution (Maybe QueryExecutionStatistics)

The amount of data scanned during the query execution and the amount of time that it took to execute.

qeQueryExecutionId :: Lens' QueryExecution (Maybe Text)

The unique identifier for each query execution.

QueryExecutionContext

data QueryExecutionContext

The database in which the query execution occurs.

See: queryExecutionContext smart constructor.

queryExecutionContext :: QueryExecutionContext

Creates a value of QueryExecutionContext with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qecDatabase - The name of the database.

qecDatabase :: Lens' QueryExecutionContext (Maybe Text)

The name of the database.

QueryExecutionStatistics

data QueryExecutionStatistics

The amount of data scanned during the query execution and the amount of time that it took to execute.

See: queryExecutionStatistics smart constructor.

queryExecutionStatistics :: QueryExecutionStatistics

Creates a value of QueryExecutionStatistics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qesEngineExecutionTimeInMillis - The number of milliseconds that the query took to execute.

qesDataScannedInBytes - The number of bytes in the data that was queried.

qesEngineExecutionTimeInMillis :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of milliseconds that the query took to execute.

qesDataScannedInBytes :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of bytes in the data that was queried.

QueryExecutionStatus

data QueryExecutionStatus

The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

See: queryExecutionStatus smart constructor.

queryExecutionStatus :: QueryExecutionStatus

Creates a value of QueryExecutionStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qesState - The state of query execution. SUBMITTED indicates that the query is queued for execution. RUNNING indicates that the query is scanning data and returning results. SUCCEEDED indicates that the query completed without error. FAILED indicates that the query experienced an error and did not complete processing. CANCELLED indicates that user input interrupted query execution.

qesStateChangeReason - Further detail about the status of the query.

qesSubmissionDateTime - The date and time that the query was submitted.

qesCompletionDateTime - The date and time that the query completed.

qesState :: Lens' QueryExecutionStatus (Maybe QueryExecutionState)

The state of query execution. SUBMITTED indicates that the query is queued for execution. RUNNING indicates that the query is scanning data and returning results. SUCCEEDED indicates that the query completed without error. FAILED indicates that the query experienced an error and did not complete processing. CANCELLED indicates that user input interrupted query execution.

qesStateChangeReason :: Lens' QueryExecutionStatus (Maybe Text)

Further detail about the status of the query.

qesSubmissionDateTime :: Lens' QueryExecutionStatus (Maybe UTCTime)

The date and time that the query was submitted.

qesCompletionDateTime :: Lens' QueryExecutionStatus (Maybe UTCTime)

The date and time that the query completed.

ResultConfiguration

data ResultConfiguration

The location in Amazon S3 where query results are stored and the encryption option, if any, used for query results.

See: resultConfiguration smart constructor.

resultConfiguration

Arguments

:: Text
rcOutputLocation

-> ResultConfiguration

Creates a value of ResultConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcEncryptionConfiguration - If query results are encrypted in S3, indicates the S3 encryption option used (for example, SSE-KMS or CSE-KMS and key information.

rcOutputLocation - The location in S3 where query results are stored.

rcEncryptionConfiguration :: Lens' ResultConfiguration (Maybe EncryptionConfiguration)

If query results are encrypted in S3, indicates the S3 encryption option used (for example, SSE-KMS or CSE-KMS and key information.

rcOutputLocation :: Lens' ResultConfiguration Text

The location in S3 where query results are stored.

ResultSet

data ResultSet

The metadata and rows that comprise a query result set. The metadata describes the column structure and data types.

See: resultSet smart constructor.

resultSet :: ResultSet

Creates a value of ResultSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsRows - The rows in the table.

rsResultSetMetadata - The metadata that describes the column structure and data types of a table of query results.

rsRows :: Lens' ResultSet [Row]

The rows in the table.

rsResultSetMetadata :: Lens' ResultSet (Maybe ResultSetMetadata)

The metadata that describes the column structure and data types of a table of query results.

ResultSetMetadata

data ResultSetMetadata

The metadata that describes the column structure and data types of a table of query results.

See: resultSetMetadata smart constructor.

resultSetMetadata :: ResultSetMetadata

Creates a value of ResultSetMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsmColumnInfo - Information about the columns in a query execution result.

rsmColumnInfo :: Lens' ResultSetMetadata [ColumnInfo]

Information about the columns in a query execution result.

Row

data Row

The rows that comprise a query result table.

See: row smart constructor.

row :: Row

Creates a value of Row with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rowData - The data that populates a row in a query result table.

rowData :: Lens' Row [Datum]

The data that populates a row in a query result table.

UnprocessedNamedQueryId

data UnprocessedNamedQueryId

Information about a named query ID that could not be processed.

See: unprocessedNamedQueryId smart constructor.

unprocessedNamedQueryId :: UnprocessedNamedQueryId

Creates a value of UnprocessedNamedQueryId with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unqiNamedQueryId - The unique identifier of the named query.

unqiErrorCode - The error code returned when the processing request for the named query failed, if applicable.

unqiErrorMessage - The error message returned when the processing request for the named query failed, if applicable.

unqiNamedQueryId :: Lens' UnprocessedNamedQueryId (Maybe Text)

The unique identifier of the named query.

unqiErrorCode :: Lens' UnprocessedNamedQueryId (Maybe Text)

The error code returned when the processing request for the named query failed, if applicable.

unqiErrorMessage :: Lens' UnprocessedNamedQueryId (Maybe Text)

The error message returned when the processing request for the named query failed, if applicable.

UnprocessedQueryExecutionId

data UnprocessedQueryExecutionId

Describes a query execution that failed to process.

See: unprocessedQueryExecutionId smart constructor.

unprocessedQueryExecutionId :: UnprocessedQueryExecutionId

Creates a value of UnprocessedQueryExecutionId with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uqeiErrorCode - The error code returned when the query execution failed to process, if applicable.

uqeiQueryExecutionId - The unique identifier of the query execution.

uqeiErrorMessage - The error message returned when the query execution failed to process, if applicable.

uqeiErrorCode :: Lens' UnprocessedQueryExecutionId (Maybe Text)

The error code returned when the query execution failed to process, if applicable.

uqeiQueryExecutionId :: Lens' UnprocessedQueryExecutionId (Maybe Text)

The unique identifier of the query execution.

uqeiErrorMessage :: Lens' UnprocessedQueryExecutionId (Maybe Text)

The error message returned when the query execution failed to process, if applicable.